requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
click==8.1.3
colorama==0.4.6
Flask==1.1.2
Flask-Migrate==2.5.2
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
greenlet==2.0.0
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.17.0
Mako==1.2.3
MarkupSafe==2.1.1
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.8.6
pyrsistent==0.19.2
python-dotenv==0.15.0
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.42
typing_extensions==4.4.0
Werkzeug==1.0.1
zipp==3.10.0
